move hcaptcha to a separate DI module
