+++ app/src/main/java/com/xinto/opencord/OpenCord.kt
@@ -21,7 +21,8 @@ class OpenCord : Application() {
                 simpleAstModule,
                 viewModelModule,
                 loggerModule,
-                providerModule
+                providerModule,
+                captchaModule
             )
         }
     }
