+++ app/src/main/java/com/xinto/opencord/di/HttpModule.kt
@@ -1,8 +1,5 @@
 package com.xinto.opencord.di
 
-import android.content.Context
-import com.hcaptcha.sdk.HCaptcha
-import com.hcaptcha.sdk.HCaptchaConfig
 import com.xinto.opencord.BuildConfig
 import com.xinto.opencord.domain.manager.AccountManager
 import com.xinto.opencord.domain.provider.PropertyProvider
@@ -30,18 +27,11 @@ val HttpHeaders.XDiscordLocale: String
     get() = "X-Discord-Locale"
 
 val httpModule = module {
+
     fun provideJson(): Json {
         return Json {
             ignoreUnknownKeys = true
         }
-    }
-
-    fun provideHCaptcha(context: Context): HCaptcha {
-        val config = HCaptchaConfig.builder()
-            .siteKey("f5561ba9-8f1e-40ca-9b5b-a0b3f719ef34") // doubt this will ever change
-            .resetOnTimeout(true)
-            .build()
-        return HCaptcha.getClient(context).setup(config)
     }
 
     fun <T : HttpClientEngineConfig> HttpClientConfig<T>.installLogging(loggerDI: Logger) {
@@ -140,7 +130,6 @@ val httpModule = module {
     }
 
     single { provideJson() }
-    single { provideHCaptcha(get()) }
     single(named("auth")) { provideAuthClient(get(), get(), get(), get()) }
     single(named("api")) { provideApiClient(get(), get(), get(), get(), get()) }
     single(named("gateway")) { provideGatewayClient() }
